requirements.txt
numpy==1.21.4
pandas==1.3.4
torch==1.10.0
pytorch_lightning==1.5.3
gpytorch==1.5.1
scikit-learn==1.0.1
scipy==1.7.3
matplotlib==3.5.0
hydra-core==1.1.1
hydra-optuna-sweeper==1.1.2
omegaconf==2.1.1
category_encoders==2.3.0
catboost==1.0.3
tqdm
torchviz
dask==2022.2.0
